feat(Backup): Refactor the way backup code is structured.
